+++ CHANGELOG.md
@@ -13,6 +13,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `Channel.update` unnecessarily causing an actual name and imageURL update on 1-1 channels [#973](https://github.com/GetStream/stream-chat-swift/issues/973)
 - `Channel.membership` is `nil` for when channel is obtained from `presenter.rx.channelDidUpdate` observable
 
+### 🔄 Changed
+- HTTP requests were incorrectly logged with `error` log level, changed to `debug`
+
 # [2.6.6](https://github.com/GetStream/stream-chat-swift/releases/tag/2.6.6)
 _May 04, 2021_